power_map

+++ Nexus27/include/character.h
@@ -6,6 +6,6 @@
 #define COLOR_SIZE 2
 // 캐릭터 관련 함수들
 void customize_character(vector<GameCharacter>& characters, bool& created);
-vector<float> random_generate(int parts);
+vector<int> random_generate(int parts);
 
 #endif // CHARACTER_H
